pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "adaptive"
version = "0.1.0"
description = "Adaptive - Continuous optimization for AI agents through intelligent observability and automated prompt evolution"
readme = "README.md"
requires-python = ">=3.12"
license = {text = "MIT"}
authors = [
    {name = "Madhur Prashant"}
]
keywords = ["ai", "agents", "langchain", "observability", "adaptive", "multi-agent", "optimization"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "bedrock-agentcore>=1.0.6",
    "boto3>=1.40.67",
    "ddtrace>=2.0.0",
    "deepagents>=0.2.5",
    "google-auth>=2.29.0",
    "google-auth-oauthlib>=1.2.0",
    "httpx>=0.28.1",
    "langchain",
    "langchain-aws>=1.0.0",
    "langchain-mcp-adapters>=0.1.13",
    "langfuse>=3.9.3",
    "langgraph-checkpoint-aws>=1.0.1",
    "mlflow[mcp]>=3.5.1",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.0",
    "requests>=2.32.0",
    "tavily>=1.1.0",
    "transformers>=4.30.0",
    "zmq>=0.0.0",
]

[project.scripts]
adaptive = "adaptive.cli:main"

[project.urls]
Homepage = "https://github.com/madhurprash/adaptive"
Documentation = "https://github.com/madhurprash/adaptive#readme"
Repository = "https://github.com/madhurprash/adaptive"
Issues = "https://github.com/madhurprash/adaptive/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/adaptive"]
